[LH-46] Add Navigation Module in app #8
