+++ app/src/main/java/com/lighthouse/lingo_swap/di/DataSourceModule.kt
@@ -12,7 +12,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-class DataSourceModule {
+object DataSourceModule {
     @Provides
     @Singleton
     fun provideDrivenDataSource(@Test drivenApiService: DrivenApiService): DrivenRemoteDataSource {
