+++ app/src/main/java/com/lighthouse/lingo_swap/di/NavigationModule.kt
@@ -1,7 +1,6 @@
 package com.lighthouse.lingo_swap.di
 
-import com.lighthouse.domain.repository.HomeRepository
-import com.lighthouse.domain.usecase.GetMatchedUserUseCase
+import com.lighthouse.navigation.Navigator
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -10,10 +9,10 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-object DomainModule {
+object NavigationModule {
     @Provides
     @Singleton
-    fun provideHomeUseCase(homeRepository: HomeRepository): GetMatchedUserUseCase {
-        return GetMatchedUserUseCase(homeRepository)
+    fun provideNavigation(): Navigator {
+        return Navigator()
     }
 }
