+++ app/src/main/java/com/lighthouse/lingo_swap/di/NetModule.kt
@@ -23,7 +23,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-class NetModule {
+object NetModule {
     @Provides
     @Singleton
     @Main
